Move some helper functions to the root of the net tests crate.
